feat: add panel position in settings
